chore(all): replace netlink library for more flexibility - drop usage of github.com/vishvananda/netlink - use github.com/mdlayher/genetlink and github.com/jsimonetti/rtnetlink - these new libraries allow more control especially to tweak conntrack in the upcoming changes - change netlinker Link methods to accept a `linkIndex uint32` instead of an entire `netlink.Link` struct. - change MTU type from uint16 to uint32 and family type from int to uint8 to match netlink kernel messages - move types from internal/netlink/types.go to their respective file in internal/netlink

+++ internal/netlink/family.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 )
 
-func FamilyToString(family int) string {
+func FamilyToString(family uint8) string {
 	switch family {
 	case FamilyAll:
 		return "all"
